예외처리 추가

+++ src/main/java/com/example/springsession/service/UserService.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @Service
 @RequiredArgsConstructor
@@ -21,12 +22,15 @@ public class UserService {
         user.setName(dto.getName());
         user.setAge(dto.getAge());
         user.setIntro(dto.getIntro());
-
         return userRepository.save(user);
     }
 
     public User getUser(Long id) { // 사용자 조회
-        return userRepository.findById(id);
+        User user = userRepository.findById(id);
+        if (user == null) {
+            throw new NoSuchElementException(id + "의 유저는 존재하지않습니다.");
+        }
+        return user;
     }
 
     public List<User> getAllUser() { // 사용자 전체 조회
@@ -34,12 +38,18 @@ public class UserService {
     }
 
     public void updateUser(Long id, UserDto dto) { // 사용자 업데이트
+        User user = userRepository.findById(id);
+        if (user == null) {
+            throw new NoSuchElementException(id + "의 유저는 존재하지않습니다.");
+        }
         userRepository.update(id, dto);
     }
 
     public void deleteUser(Long id) {
+        User user = userRepository.findById(id);
+        if (user == null) {
+            throw new NoSuchElementException(id + "의 유저는 존재하지않습니다.");
+        }
         userRepository.delete(id);
     }
-
-
 }
